Create LICENCE
It is a good rule of thumb that everything should have a license so that is it clear what others can and cannot do. This is a copyleft license that facilitates the sharing of code. It is not important in this project but i have this file in all project as a principle.